reuse hyrbid highlighter in multi highglighter
